refactor: code refactor and adding style
