+++ src/App.tsx
@@ -1,7 +1,11 @@
-import GeneratePassword from './componsnts/generate-password';
+import GenerateForm from './componsnts/generate-form';
 
 function App() {
-  return <GeneratePassword />;
+  return (
+    <main className="bg-linear-to-bl h-dvh from-blue-950 via-purple-950 to-gray-900 p-4 text-white">
+      <GenerateForm />
+    </main>
+  );
 }
 
 export default App;
